trackly-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
alembic
pytesseract
Pillow
openai
xlsxwriter
reportlab
python-dotenv
PyPDF2
pandas